fixed text highlighter PHP 7.1 compatibility
for some reason $this->_output is a string at this point so it will
crash when [] is applied to it. PHP 7.1 got stricter about this and does
not just convert the whole thing to an array.
This is not the real bug fix but a workaround to mimic PHPs behaviors in
versions < 7.1